:recycle: [filestorage] Extract module `observers` from `storage`
